requirements.txt
numpy>=1.19.5
pandas>=1.3.0
matplotlib>=3.4.2
scikit-learn>=0.24.2
opencv-python>=4.5.3
insightface>=0.6.0
facenet-pytorch>=2.5.2
torch>=1.9.0
torchvision>=0.10.0
tqdm>=4.62.0
seaborn>=0.11.1
scipy>=1.7.0
pillow>=8.3.1
kaggle